# specs/api.openapi.yaml
components:
  schemas:
...
    Body_chat_voice_endpoint_api_chat_voice_post:
      properties:
        agent:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent
        audio:
          contentMediaType: application/octet-stream
          title: Audio
          type: string
        lang:
          anyOf:
          - type: string
          - type: 'null'
          title: Lang
        session_id:
          default: default
          title: Session Id
          type: string
      required:
      - audio
      title: Body_chat_voice_endpoint_api_chat_voice_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: LLM agent orchestration service
  title: Helperium API
  version: 1.1.0
openapi: 3.1.0
paths:
...
  /api/chat/voice:
    post:
      operationId: chat_voice_endpoint_api_chat_voice_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_chat_voice_endpoint_api_chat_voice_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Chat Voice Endpoint
  /api/chat/{name}:
    post:
      operationId: chat_agent_handler_api_chat__name__post
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Chat Agent Handler
...